Extract use-control-point.tsx
