+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/movable-line.tsx
@@ -1,5 +1,5 @@
 import {vec} from "mafs";
-import {useRef, useState} from "react";
+import {useRef} from "react";
 import * as React from "react";
 
 import {inset, snap, size} from "../../math";
@@ -9,8 +9,8 @@ import {useDraggable} from "../use-draggable";
 import {useTransformVectorsToPixels} from "../use-transform";
 import {getIntersectionOfRayWithBox} from "../utils";
 
-import {MovablePointView} from "./movable-point-view";
 import {SVGLine} from "./svg-line";
+import {useControlPoint} from "./use-control-point";
 import {Vector} from "./vector";
 
 import type {Interval} from "mafs";
@@ -72,55 +72,6 @@ export const MovableLine = (props: Props) => {
         </>
     );
 };
-
-function useControlPoint(
-    point: vec.Vector2,
-    color: string | undefined,
-    onMovePoint: (newPoint: vec.Vector2) => unknown,
-) {
-    const {snapStep, disableKeyboardInteraction} = useGraphConfig();
-    const [focused, setFocused] = useState(false);
-    const keyboardHandleRef = useRef<SVGGElement>(null);
-    useDraggable({
-        gestureTarget: keyboardHandleRef,
-        point,
-        onMove: onMovePoint,
-        constrainKeyboardMovement: (p) => snap(snapStep, p),
-    });
-
-    const visiblePointRef = useRef<SVGGElement>(null);
-    const {dragging} = useDraggable({
-        gestureTarget: visiblePointRef,
-        point,
-        onMove: onMovePoint,
-        constrainKeyboardMovement: (p) => snap(snapStep, p),
-    });
-
-    const focusableHandle = (
-        <g
-            data-testid="movable-point__focusable-handle"
-            className="movable-point__focusable-handle"
-            tabIndex={disableKeyboardInteraction ? -1 : 0}
-            ref={keyboardHandleRef}
-            onFocus={() => setFocused(true)}
-            onBlur={() => setFocused(false)}
-        />
-    );
-    const visiblePoint = (
-        <MovablePointView
-            point={point}
-            dragging={dragging}
-            color={color}
-            ref={visiblePointRef}
-            focusBehavior={{type: "controlled", showFocusRing: focused}}
-        />
-    );
-
-    return {
-        focusableHandle,
-        visiblePoint,
-    };
-}
 
 const defaultStroke = "var(--movable-line-stroke-color)";
